requirements.txt
aws-cdk-lib==2.250.0
constructs>=10.0.0,<11.0.0
boto3==1.42.89
aws-lambda-powertools>=3.18